7/26/15
HUZZAH - Got the last bit of functionality work.
- Categories work(can add category to a post, can display the category
on the index page, the listed category is a link to show all posts
within that category)
- Created a partial view that works as the Sidebar for both Index and
Details.  It has the search bar at the top, a list of blog categories in
the middle (clicking on any of the category links now takes you to a
list of posts within that category), and a User panel at the bottom
(this is where I keep the login/logout/admin panel links)